2009-11-21  Jean Brefort  <jean.brefort@normalesup.org> 	* configure.ac: request goffice >= 0.7.16. 	* libs/gcu/cmd-context.cc: fixed after goffice API change.

+++ libs/gcu/cmd-context.cc
@@ -54,7 +54,7 @@ gcu_cc_set_sensitive (G_GNUC_UNUSED GOCmdContext *cc,
 }
 
 static void
-gcu_cc_progress_set (G_GNUC_UNUSED GOCmdContext *cc, G_GNUC_UNUSED gfloat val)
+gcu_cc_progress_set (G_GNUC_UNUSED GOCmdContext *cc, G_GNUC_UNUSED double val)
 {
 }
